Add simple viewport mode cycling
The text view I'd meant to work on doesn't make sense as part of this,
since cycling with Tab will just type a tab character in a TextArea,
which means you can't get out without using the drop down, which is a
pain. A separate window or panel for text would be better.