WIP: Clean up texture coordinate code
I think, in my quest to make the coords 0 through 1, instead of allowing negative and/or further from zero than 1, I might have ultimately come full circle and not changed anything, to be honest, but the code is cleaner, at least. I'll be squashing a bunch of these eventually, so who cares? Oh, and the big thing seems to have been setting the texture mode to repeat instead of clamp.