Switch to custom OpenTK NuGet package
It pains me to strip out a demonstration of Nuke's flexibility, but the burden of rebuilding the custom OpenTK was too much. Not just the 10+ minutes of wasted time, but the CPU cycles, RAM, disk space, electricity and A/C consumed by my free Azure Pipelines account gnawed at my conscience.